% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{KMEANS}
\alias{KMEANS}
\title{KMEANS}
\usage{
KMEANS(data_matrix, K)
}
\arguments{
\item{data_matrix}{An N x M matrix, containing N rows (observations) and M columns (data features)}

\item{K}{A positive integer describing the number of clusters the clustering should produce.}
}
\value{
A list containing the following results:

\code{cluster}: a vector of length N containing the cluster ids of each observation. Cluster ids are positive integers from 1:K inclusive.

\code{centers}: a matrix of cluster centers, with K rows, and M columns.

\code{withinss}: a vector of length K, containing within-cluster sums of squares for each cluster.

\code{tot.withinss}: the total within-cluster sum of squares

\code{size}: a vector of length K, containing the number of points in each cluster.
}
\description{
Generates a K-means clustering of a given feature matrix, with K cluster centers.
K-means is a common non-probabilistic clustering method, useful when data can be
fit to circular clusters.
}
\examples{
# Cluster the iris data set into three clusters:
KMEANS(iris[1:4], K=3)
}
\seealso{
\href{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/kmeans}{stats::kmeans} for a more robust implementation.
}
